requirements.txt
mysql-connector-python

# the command to download everything 
# pip install -r requirements.txt